backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
openai==1.54.0
anthropic==0.39.0
psycopg2-binary==2.9.10
pgvector==0.3.5
redis==5.2.0
websockets==13.1
pydantic-settings==2.6.0
PyJWT==2.9.0
